+++ src/components/ui/Spinner/index.module.scss
