+++ src/layouts/CategoryResult/SearchContents/ProductTab/index.module.scss
